Improved the way the DbConnection is configured.
Before, the Database constructor accepted a Func<DbConnection> which it
will use to create the connections when needed. However, this was rather
inconvenient and ugly. It also made it impossible to use the specific
connection type without down-casting, which is worse than ugly.
This new version uses Generics instead. Configuring the DbConnection is
just a matter of specifying the correct type when calling the constructor
and now ConnectionDo(..) will accept a type-safe Func or Action.