Big change to implement transaction support.
As discussed with Tera, I added a DatabaseContext class, which is the
superclass of both Database and the newly added DatabaseTransaction. The
DatabaseContext public interface consists only of the methods Query(..) and
NonQuery(..).
Database now also includes a TransactionDo(..) method which accepts an
action whose only parameter is an instance of DatabaseTransaction. If this
action gets evaluated and no DbException error is thrown, the transaction
gets committed, otherwise it gets rolled back and the exception is thrown
up the call stack.